# feat: faster IndexedCrate::new with rayon