--- a/Cargo.toml
+++ b/Cargo.toml
@@ -12,7 +12,15 @@ readme = "./README.md"
 
 [dependencies]
 trustfall = "0.7.1"
+rayon = { version = "1.10.0", optional = true }
 rustdoc-types = "0.29.1"
+
+[features]
+default = []
+# Enable rayon for speeding up building `IndexedCrate`, this massively improves
+# performance on big crates, but may impact performance on small crates and
+# definitely increases the memory usage
+rayon = ["dep:rayon"]
 
 [[bench]]
 name = "indexed_crate"